mark edge of bin as within centromere or not

Can optionally specify a padding to mark locations of bins as being close enough to a centromere. Often bins near centromeres are corrupt in their state calls. In the past, we have filtered within 3 Mb of centromeres.

Usage

bin_within_centromere(
  reads_df,
  padding = 0,
  bin_location_column = "start",
  hg19 = FALSE,
  hg38 = FALSE
)

Arguments

reads_df

tibble of read data

padding

int of number of BP to add to each side of the centromere

hg19

bool. Not needed unless this function is also adding the centromere information to the data frame. See add_centromere_locations() which will be called if the correct centromere columns are not present.

hg38

bool. Same as hg19...but for hg38.

bin_locatation_column

which column to use as the location of a bin. popular choices are the start of a bin, middle of a bin, or and of a bin.

Value

input table, but with a boolean 'within_centro' column added (and potentially other centromere information columns, if needed)

Details

IMPORTANT! This adds the padding to each side of the centromere. So if you specify a padding of 3 Mb, it will be within 3 Mb of the start and 3 Mb of the end of the centromere.